Rename BaseMappedList to TypeScript
